Implement roles
Add 'Role' table with unique names for roles. Add 'Permissions' table to
associate a timespan with a role.

TL;DR: Roles lets us manage timespan permissions.

The basic idea of how we envision roles being used is a bit like UNIX
groups, in that every user gets a role with the same name as their
username. They are also added to a User role. An administrator or
moderator is further added to a Moderator role or equivalent. The whole
idea does go well beyond UNIX groups in that a timespan has per-role
permissions. Again, none of this is enforced by tempuhs. It's
strictly for the levels above tempuhs to decide.

There are four types of permissions: own, read, write and share
permissions. What these mean is also up to the implementation. The idea
is that owners get to change permissions, readers get to access and view
the timespans, writers get to change the timespans, and sharers get to
make clones of the timespans for their own timelines.

To settle what permissions a user has for a given timespan, given that
said user will likely have several roles that might have conflicting
permissions set for the timespan, the permissions are summed. I.e. if
any of the roles gives the user permission to read/write/share, the user
is permitted to do so.

Implementations of a tempuhs-based system might want to have a
friends-system. The way to do this would be to make a role for a user's
friends, and let their friends take on this role. As an example,
consider user Foo who wants to add friends Bar and Baz. The role
Foo_friends is made, and Bar and Baz are given this role.

There need to be rules in place for permissions of new timespans. A
suggestion would be to have all users in a User role, and then set own
to the current user's personal role, and read/write/share to 'True' or
'False', or something more complex, depending on which makes sense in
your tempuhs-based implementation.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen <stian@plaimi.net>
Reviewed-by: Stian Ellingsen <stian@plaimi.net>