Restructure code into useful modules
Put types in a 'Types' module and functions in a 'Functions' module, and
database stuff in a 'Database' module. This makes importing stuff more
flexible for library users. It also makes it easier to figure out where
stuff is and what's related to what.

Export everything useful in the Chronology module as before.

This change is also partly motivated by isolating TH.

In practice this does not change anything right now for users that
import 'Tempuhs.Chronology', but they may now import only the useful
modules instead of the whole thing.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen <stian@plaimi.net>
Reviewed-by: Stian Ellingsen <stian@plaimi.net>